% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteristics.R
\name{characteristics}
\alias{characteristics}
\title{balancedSurveyR: Examine sample/population characteristics}
\usage{
characteristics(sample, population, model, state = "sampled",
  plot_type = "density")
}
\arguments{
\item{sample}{(dataframe) Your sample returned by sampler()}

\item{population}{(dataframe) Your population dataframe sent to sampler()}

\item{model}{(balancedSurveyR model object) the output of a historical_model() fit your historical data.}

\item{state}{(character) Either 'sampled' or 'responded' (default: 'sampled')}

\item{plot_type}{(character) Either 'density' or 'ecdf' (default: density)}
}
\value{
ggplot2 object
}
\description{
balancedSurveyR: Examine sample/population characteristics
}
\author{
Brock Ferguson, \email{brock.ferguson@gmail.com}

Examine sample characteristics before/after responses

After generating a new sample, you can examine the performance
of this tool by examining the characteristics of your samples.

When called with state='sampled', you'll view the characteristics
of your sample in its entirety. This should over-represent segments
of your population that are unlikely to respond, because we need to
ask more of them to complete the survey than your typical person.

When called with state='responded', you'll view the expected characteristics
of the portion of your sample that responds to the survey. Note: this is based
off their *expected* response rates: if people don't respond in the future
like they have to this point, then this may be inaccurate.
}

