% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_model.R
\name{historical_model}
\alias{historical_model}
\title{balancedSurveyR: Historical model}
\usage{
historical_model(data, attribute_columns, response_column, model = "xgbTree")
}
\arguments{
\item{data}{(dataframe/tbl) of historical response data.}

\item{attribute_columns}{(vector) column names indicating possibly-predictive attributes}

\item{response_column}{(character) column name for response (1/0) column}

\item{model}{(character) type of model to use on the data}
}
\value{
balancedSurveyR model object, a list with keys: model, fit, predictions
}
\description{
balancedSurveyR: Historical model
}
\examples{
library('dplyr')
library('ggplot2')

# generate fake historical data
historical_responses <- expand.grid(
  person=1:100,
  age = 20:50
) \%>\%
  mutate(
    responded = rbinom(nrow(.), size=1, prob=(age / 100))
  ) \%>\%
  select(-starts_with('.')) \%>\%
  select(-person)

# fit historical_model()
model <- historical_model(data = historical_responses,
                          attribute_columns = c('age'),
                          response_column = 'responded',
                          model = 'xgbTree')

# verify predictions
historical_responses$predictions <- model$predictions
ggplot(historical_responses, aes(x=predictions, y=responded)) +
  stat_summary(fun.y='mean', geom='point') +
  stat_smooth(method="glm", method.args=list(family="binomial"))

# generate new population data
survey_population <- expand.grid(
  person=1:1000,
  age = 20:50
) \%>\%
  select(-person)

# generate sample
sampled_population <- sampler(population = survey_population,
                              model = model,
                              desired_n = 1000)

# examine characteristics
# sampled characteristics should be biased towards under-responders
sampled_characteristics <- characteristics(sample = sampled_population,
                                           population = survey_population,
                                           model = model,
                                           state = 'sampled')

sampled_characteristics

# responder characteristics should be representative of population
responder_characteristics <- characteristics(sample = sampled_population,
                                             population = survey_population,
                                             model = model,
                                             state = 'responded')

responder_characteristics

}
\author{
Brock Ferguson, \email{brock.ferguson@gmail.com}

Fit a model to historical response data

This should be the first function you call when using balancedSurveyR.
You will feed it a dataframe of existing survey records. Each record
should have 1+ columns of "attributes" (things you think might matter
in responding to a survey) and a column indicating whether they
responded or not (1/0).

This function will then model these data using the model indicated
(default: 'xgbTree') and then use a smoothing spline to recover
best possible estimates of response rates for each prediction.

It will return a balancedSurveyR model object which can be passed
to the sample() function.
}

