% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sampler}
\alias{sampler}
\title{balancedSurveyR: Gather new sample}
\usage{
sampler(population, model, desired_n)
}
\arguments{
\item{population}{(dataframe/tbl) new people to consider surveying (i.e., your population).}

\item{model}{(balancedSurveyR model object) the output of a historical_model() fit your historical data.}

\item{desired_n}{(integer) how many people do you want to *respond* to your survey (your desired N)?}
}
\value{
dataframe of sampled clients
}
\description{
balancedSurveyR: Gather new sample
}
\author{
Brock Ferguson, \email{brock.ferguson@gmail.com}

Gather a new, balanced sample

This should be the second function you call when using balancedSampler.
After you generate your historical_model(), you will feed in a new dataframe
to this function. It will return to you a dataframe representing unique
sample of people that, when surveyed, should generate a representative
sample of your population approximately the size of your desired N(umber)
of responses.

This balanced sample is gathered through the following process:

1. Fit the historical model to the new population
2. Weight people in the new sample as the inverse of their response probability
   (w = 1 / prob).
3. Repeatedly draw weighted samples of clients until the sum of their response probabilities
   is greater or equal to the desired N.
4. Return the compiled sample as the suggested sample.
}

